backend/InformingChoicesBackend/requirements.txt
django>=3.2.14
djangorestframework>=3.16.1